# minimize HTTP/2 request latency