--- a/include/h2o/socket.h
+++ b/include/h2o/socket.h
@@ -110,6 +110,7 @@ struct st_h2o_socket_t {
     struct st_h2o_socket_peername_t *_peername;
     struct {
         uint8_t state; /* one of H2O_SOCKET_LATENCY_STATE_* */
+        uint8_t notsent_is_minimized : 1;
         uint16_t suggested_tls_payload_size;
         size_t suggested_write_size; /* SIZE_MAX if no need to optimize for latency */
     } _latency_optimization;