--- a/lib/http2/scheduler.c
+++ b/lib/http2/scheduler.c
@@ -357,3 +357,10 @@ int h2o_http2_scheduler_run(h2o_http2_scheduler_node_t *root, h2o_http2_schedule
     }
     return 0;
 }
+
+int h2o_http2_scheduler_is_active(h2o_http2_scheduler_node_t *root)
+{
+    return root->_queue != NULL && root->_queue->bits != 0;
+}
+
+