WIP.

The basic idea is to only write(2) data that can be immediately sent by the TCP/IP stack, by using TCP_NOTSENT_LOWAT and TCP_INFO.