@kazuho 2016-05-13
@deweerdt 
> My reading of the `TCP_NOTSENT_LOWAT` patch (https://lwn.net/Articles/560082/) is that setting 1 as a value means that `epoll()` will only call `EPOLLOUT` if there are `0` unsent bytes on the socket. Is that the intended behavior?

Yes.

Eric Dumazet and we have different motivations.  Eric's intention is to reduce memory usage without sacrificing throughput.  In our case, we would like to optimize for latency when RTT is high and CWND is small, so that the server can start responding to a new request immediately with a higher priority.

When we are to optimize for latency, we would want to instruct `epoll` to notify us only when we can send more than `1` byte immediately using the socket (which would naturally mean `TCP_NOTSENT_LOWAT = 0`).  But linux seems to not support such a setting (or it is broken).  Therefore, we resort to using `1` as the argument.

> I also believe that setting the value to `0` won't disable the feature, but will make it use the system wide value in `/proc/sys/net/ipv4/tcp_notsent_lowat`

It is true that setting `TCP_NOTSENT_LOWAT` via `setsockopt` will make the socket use the system wide value.  OTOH, you can set `/proc/sys/net/ipv4/tcp_notsent_lowat` to zero (note the default is `SIZE_MAX`).  However setting it to zero breaks existing applications; for example, sshd stops responding.  I assume this to be a bug.