@deweerdt 2016-05-13
Hello @kazuho,

Thank you for response, I belive I understand the goal of the PR
well. I'm however still confused by the change, I hope you'll forgive
me for being dense.

From `tcp_poll`, the kernel will notify us if sk_stream_is_writeable()
returns true:
```
    if (sk_stream_is_writeable(sk)) {
        mask |= POLLOUT | POLLWRNORM;
        ...
```


In `sk_stream_is_writeable`, the first element looks for space in
the send buffer and calls `sk_stream_memory_free`.

```
static inline bool sk_stream_is_writeable(const struct sock *sk)
{
        return sk_stream_wspace(sk) >= sk_stream_min_wspace(sk) &&
               sk_stream_memory_free(sk);
}
```

`sk_stream_memory_free` then calls `tcp_stream_memory_free`. And
that will only return true if the bytes not yet sent are lower than
the `TCP_NOTSENT_LOWAT` value.

So if we set `TCP_NOTSENT_LOWAT` to `1`, if there's space in the
send buffer, but there are 2 bytes that we have not sent, we will not be notified,
when the write would have in fact succeeded, had we left
`TCP_NOTSENT_LOWAT` to its default `0xffffffff` value.

```
static inline u32 tcp_notsent_lowat(const struct tcp_sock *tp)
{
        struct net *net = sock_net((struct sock *)tp);
        return tp->notsent_lowat ?: net->ipv4.sysctl_tcp_notsent_lowat;
}

static inline bool tcp_stream_memory_free(const struct sock *sk)
{
        const struct tcp_sock *tp = tcp_sk(sk);
        u32 notsent_bytes = tp->write_seq - tp->snd_nxt;

        return notsent_bytes < tcp_notsent_lowat(tp);
}
```

By setting `TCP_NOTSENT_LOWAT`, I believe we will use very little
kernel memory, but we will increase the risk that the send buffer
will be empty when it would have been possible to send data.

> OTOH, you can set /proc/sys/net/ipv4/tcp_notsent_lowat to zero (note the default is SIZE_MAX). However setting it to zero breaks existing applications; for example, sshd stops responding. I assume this to be a bug.

If my understanding is correct, setting it to 0 would mean that we never are notified by `epoll`, since `notsent_bytes < 0` can never be true. That would be a kernel bug and the sysctl shouldn't accept a value of `0`.

What am I missing?
Thanks,
Frederik