@kazuho 2016-05-13
@deweerdt Thank you for looking into this.  It's complicated, and I am happy to have your eyes checking what I am doing.

> So if we set `TCP_NOTSENT_LOWAT` to `1`, if there's space in the send buffer, but there are 2 bytes that we have not sent, we will not be notified, when the write would have in fact succeeded, had we left `TCP_NOTSENT_LOWAT` to its default `0xffffffff` value.

I understand the same.

> By setting `TCP_NOTSENT_LOWAT`, I believe we will use very little kernel memory, but we will increase the risk that the send buffer will be empty when it would have been possible to send data.

Correct.

The bandwidth (per TCP connection) we are going to loose in case of `TCP_NOTSENT_LOWAT` being set to a very small value depends on the ratio between the latency (at the TCP level) and the time spent for a single event loop.  If the TCP latency is much greater than the time spent for a single event loop, the issue would not matter a lot.  And fortunately, we do not need this kind of optimization for low-latency connections.

Therefore, this PR activates the optimization only if there is certain amount of latency.  I believe it is hard-coded to 20ms (in the assumption that a single event loop would spend much less time than that), but it should better be configurable.  Or we should observe the time spent for the event loop, and use the value to decide whether if we should turn on this optimization.

> If my understanding is correct, setting it to 0 would mean that we never are notified by `epoll`, since `notsent_bytes < 0` can never be true. That would be a kernel bug and the sysctl shouldn't accept a value of `0`.

Interesting.  I had never thought it that way.

My view on the parameter is that we have a _single_ buffer that stores unacked bytes _and_ unsent bytes.  Therefore, to me `notsent_bytes = 0` is a valid configuration.