@deweerdt 2016-05-14
Hello @kazuho 

> The bandwidth (per TCP connection) we are going to loose in case of TCP_NOTSENT_LOWAT being set to a very small value depends on the ratio between the latency (at the TCP level) and the time spent for a single event loop. If the TCP latency is much greater than the time spent for a single event loop, the issue would not matter a lot. And fortunately, we do not need this kind of optimization for low-latency connections.

> Therefore, this PR activates the optimization only if there is certain amount of latency. I believe it is hard-coded to 20ms (in the assumption that a single event loop would spend much less time than that), but it should better be configurable. Or we should observe the time spent for the event loop, and use the value to decide whether if we should turn on this optimization.

I think that makes sense, I hadn't considered that particular use case: basically we lower the responsiveness towards high-latency sockets in order to improve the general latency. That's pretty cool, thanks for bearing with me.

> > If my understanding is correct, setting it to 0 would mean that we never are notified by epoll, since notsent_bytes < 0 can never be true. That would be a kernel bug and the sysctl shouldn't accept a value of 0.
> Interesting. I had never thought it that way.

> My view on the parameter is that we have a single buffer that stores unacked bytes and unsent bytes. Therefore, to me notsent_bytes = 0 is a valid configuration.

Maybe I should send an email to the lkml? That would help clarify the case of the sysctl set to `0`.