@kazuho 2016-05-16
At e4f89d4 with some debug logs added, I run a preliminary bencmark using [latopt.tar.gz](https://github.com/h2o/h2o/files/265686/latopt.tar.gz).

On a 250ms RTT network, the latency for obtaining a JS referred from a large HTML decreases to about half (from 500ms to 250ms).
![latopt-chrome-ireland-japan](https://cloud.githubusercontent.com/assets/41567/15282389/f1fb2a94-1b7d-11e6-95b3-fa2d5652ef5a.gif)

On a 25ms RTT network, the latency goes down from 50ms to 33ms (for this benchmark, I run several times to get the average number).
![latopt-25ms](https://cloud.githubusercontent.com/assets/41567/15282390/f4b42484-1b7d-11e6-8a2c-f57136297f70.png)